wmaker: fix non-portable int argument for printf in error message
The original code assumed that the sizeof returns a long int type, which is
true on 64 bits platform but not on 32 bits platforms, as pointed by gcc.

The new code does an explicit conversions to 'int' (which should be enough
for that case) so the format specifier in printf will always be ok,
whatever platform Window Maker is compiled for.

+++ src/dialog.c
@@ -591,8 +591,8 @@ static void listPixmaps(WScreen *scr, WMList *lPtr, const char *path)
 		if (wstrlcpy(pbuf, apath, sizeof(pbuf)) >= sizeof(pbuf) ||
 		    wstrlcat(pbuf, "/", sizeof(pbuf)) >= sizeof(pbuf) ||
 		    wstrlcat(pbuf, dentry->d_name, sizeof(pbuf)) >= sizeof(pbuf)) {
-			wwarning(_("full path for file \"%s\" in \"%s\" is longer than %ld bytes, skipped"),
-			         dentry->d_name, path, sizeof(pbuf) - 1);
+			wwarning(_("full path for file \"%s\" in \"%s\" is longer than %d bytes, skipped"),
+			         dentry->d_name, path, (int) (sizeof(pbuf) - 1) );
 			continue;
 		}
